# docs/openapi.yaml
openapi: 3.0.0
paths:
  /device-tokens:
    post:
      operationId: DeviceTokenController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceTokenDto'
      responses:
        '201':
          description: 生成 device token(返回明文,供 SDK 配置)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 生成 device token(返回明文,供 SDK 配置)
      tags:
        - device-token
      description: 生成 device token(返回明文,供 SDK 配置)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    get:
      operationId: DeviceTokenController_list
      parameters: []
      responses:
        '200':
          description: 列表:所有 device token(含明文、project 名、在线设备数)成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceTokenRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 列表:所有 device token(含明文、project 名、在线设备数)
      tags:
        - device-token
      description: 列表:所有 device token(含明文、project 名、在线设备数)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /device-tokens/{id}/projects:
    patch:
      operationId: DeviceTokenController_updateProjects
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceTokenProjectsDto'
      responses:
        '200':
          description: 替换 device token 的功能组作用域成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 替换 device token 的功能组作用域
      tags:
        - device-token
      description: 替换 device token 的功能组作用域。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /device-tokens/{id}/revoke:
    post:
      operationId: DeviceTokenController_revoke
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '201':
          description: 撤销 device token成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 撤销 device token
      tags:
        - device-token
      description: 撤销 device token。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /device-tokens/{id}:
    delete:
      operationId: DeviceTokenController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 删除 device token(软删,与撤销正交)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 删除 device token(软删,与撤销正交)
      tags:
        - device-token
      description: 删除 device token(软删,与撤销正交)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /projects:
    get:
      operationId: ProjectsController_list
      parameters: []
      responses:
        '200':
          description: 功能组列表成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 功能组列表
      tags:
        - projects
      description: 功能组列表。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    post:
      operationId: ProjectsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectDto'
      responses:
        '201':
          description: 创建功能组成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 创建功能组
      tags:
        - projects
      description: 创建功能组。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /projects/info:
    get:
      operationId: ProjectsController_info
      parameters: []
      responses:
        '200':
          description: 功能组派生统计(设备数/在线/近7天/成功率/运行态)成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 功能组派生统计(设备数/在线/近7天/成功率/运行态)
      tags:
        - projects
      description: 功能组派生统计(设备数/在线/近7天/成功率/运行态)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /projects/{id}:
    delete:
      operationId: ProjectsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 删除功能组成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 删除功能组
      tags:
        - projects
      description: 删除功能组。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /projects/{id}/enabled:
    post:
      operationId: ProjectsController_setEnabled
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetEnabledDto'
      responses:
        '201':
          description: 启用/停用功能组(停用后 invoke 拒派)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectEnabledResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 启用/停用功能组(停用后 invoke 拒派)
      tags:
        - projects
      description: 启用/停用功能组(停用后 invoke 拒派)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
...
  /devices/{id}:
    get:
      operationId: DevicesController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 设备详情成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 设备详情
      tags:
        - devices
      description: 设备详情。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
...
  /monitor/requests:
    get:
      operationId: MonitorController_list
      parameters:
        - name: project
          required: false
          in: query
          schema:
            type: string
        - name: action
          required: false
          in: query
          schema:
            type: string
        - name: clientId
          required: false
          in: query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
        - name: payloadState
          required: false
          in: query
          schema:
            type: string
        - name: minimumLatencyMs
          required: false
          in: query
          schema:
            type: number
        - name: maximumLatencyMs
          required: false
          in: query
          schema:
            type: number
        - name: from
          required: false
          in: query
          description: ISO 起始时间
          schema:
            type: string
        - name: to
          required: false
          in: query
          description: ISO 结束时间
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 10
            type: number
      responses:
        '200':
          description: 请求记录列表(查 PG 脊柱,不返 payload)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestLogPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 请求记录列表(查 PG 脊柱,不返 payload)
      tags:
        - monitor
      description: 请求记录列表(查 PG 脊柱,不返 payload)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /monitor/request-options:
    get:
      operationId: MonitorController_requestOptions
      parameters:
        - name: project
          required: false
          in: query
          schema:
            type: string
        - name: action
          required: false
          in: query
          schema:
            type: string
        - name: clientId
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 请求筛选下拉选项(去重 project/action/client,联动过滤 + 实体仍存在)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestOptions'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 请求筛选下拉选项(去重 project/action/client,联动过滤 + 实体仍存在)
      tags:
        - monitor
      description: 请求筛选下拉选项(去重 project/action/client,联动过滤 + 实体仍存在)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /monitor/requests/{requestId}:
    get:
      operationId: MonitorController_detail
      parameters:
        - name: requestId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 请求详情(PG 脊柱 + 懒加载 Manticore payload/AppAudit)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 请求详情(PG 脊柱 + 懒加载 Manticore payload/AppAudit)
      tags:
        - monitor
      description: 请求详情(PG 脊柱 + 懒加载 Manticore payload/AppAudit)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
...
info:
  title: R2RPC API
  description: R2RPC 管理控制面、调用方 RPC 与监控查询的 HTTP API。设备长连接协议使用 WebSocket，另见项目协议文档。
  version: 0.1.0
  contact:
    name: R2RPC Contributors
    url: https://github.com/rezoch340/R2RPC
    email: ''
  license:
    name: UNLICENSED
    url: https://github.com/rezoch340/R2RPC/blob/main/LICENSE
...
servers:
  - url: /
    description: 当前部署
components:
  securitySchemes:
    adminJwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: 后台登录接口签发的管理员 JWT。
    accessToken:
      scheme: bearer
      bearerFormat: R2RPC Access Token
      type: http
      description: 访问令牌管理页面生成的 rk_ 前缀 Access Token。
  schemas:
    CreateDeviceTokenDto:
      type: object
      properties:
        name:
          type: string
        projects:
          type: array
          items:
            type: string
        expiresAt:
          type: string
        description:
          type: string
      required:
        - name
        - projects
    UpdateDeviceTokenProjectsDto:
      type: object
      properties:
        projects:
          description: 更新后的完整功能组名称列表
          type: array
          items:
            type: string
      required:
        - projects
    CreateProjectDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 255
      required:
        - name
    SetEnabledDto:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - timestamp
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
              example: 请求参数非法
            - type: array
              items:
                type: string
              example:
                - name must be a string
            - type: object
              additionalProperties: true
              description: Nest HttpException 的结构化响应，例如 statusCode/message/error。
        timestamp:
          type: string
          format: date-time
    DeletedResult:
      type: object
      required:
        - deleted
      properties:
        deleted:
          type: boolean
          enum:
            - true
...
    DeviceTokenRecord:
      type: object
      required:
        - id
        - name
        - token
        - status
        - createdAt
        - projects
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: production-caller
        token:
          type: string
          description: 明文令牌；Access Token 使用 rk_，Device Token 使用 dk_。
        status:
          type: string
          enum:
            - active
            - revoked
        expiresAt:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
          nullable: true
        createdBy:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        projects:
          type: array
          items:
            type: string
        onlineDeviceCount:
          type: integer
          minimum: 0
          description: 列表接口返回的当前在线设备数
    Project:
      type: object
      required:
        - id
        - name
        - enabled
        - createdAt
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: cn-nodes
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
    ProjectEnabledResult:
      type: object
      required:
        - id
        - name
        - enabled
      properties:
        id:
          type: integer
        name:
          type: string
        enabled:
          type: boolean
    ProjectInfo:
      type: object
      required:
        - id
        - name
        - enabled
        - totalDevices
        - onlineDevices
        - requests7d
        - success7d
        - successRate
        - status
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
        totalDevices:
          type: integer
          minimum: 0
        onlineDevices:
          type: integer
          minimum: 0
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        requests7d:
          type: integer
          minimum: 0
        success7d:
          type: integer
          minimum: 0
        successRate:
          type: number
          minimum: 0
          maximum: 100
        status:
          type: string
          enum:
            - online
            - offline
            - stale
            - disabled
            - no_device
    Device:
      type: object
      required:
        - id
        - clientId
        - online
        - status
      properties:
        id:
          type: integer
        clientId:
          type: string
        deviceTokenId:
          type: integer
          nullable: true
        online:
          type: boolean
        status:
          type: string
          enum:
            - online
            - offline
            - stale
        platform:
          type: string
          nullable: true
        lastIp:
          type: string
          nullable: true
        extra:
          type: string
          nullable: true
        maxInFlight:
          type: integer
          nullable: true
          minimum: 1
        description:
          type: string
          nullable: true
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    RequestLog:
      type: object
      required:
        - id
        - requestId
        - projectName
        - actionName
        - status
        - payloadState
        - createdAt
      properties:
        id:
          type: integer
          example: 1
        requestId:
          type: string
          format: uuid
        projectName:
          type: string
          example: cn-nodes
        actionName:
          type: string
          example: ping
        clientId:
          type: string
          nullable: true
        requesterUserId:
          type: integer
          nullable: true
        accessTokenId:
          type: integer
          nullable: true
        status:
          type: string
          example: ok
          description: 业务状态，例如 ok、timeout、error、no_project、no_device、offline、disabled、rejected
        httpCode:
          type: integer
          nullable: true
          example: 200
        latencyMs:
          type: integer
          nullable: true
          minimum: 0
        errorMessage:
          type: string
          nullable: true
        payloadState:
          type: string
          enum:
            - pending
            - indexed
            - failed
            - unavailable
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
    RequestLogPage:
      type: object
      required:
        - rows
        - page
        - pageSize
        - total
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/RequestLog'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
    RequestOptions:
      type: object
      required:
        - projects
        - actions
        - clientIds
      properties:
        projects:
          type: array
          items:
            type: string
          maxItems: 200
        actions:
          type: array
          items:
            type: string
          maxItems: 200
        clientIds:
          type: array
          items:
            type: string
          maxItems: 200
    AppAuditMetadata:
      type: object
      additionalProperties: false
      required:
        - key
        - value
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 100
        value:
          description: 任意可 JSON 序列化的值
          nullable: true
    AppAuditRequest:
      type: object
      additionalProperties: false
      properties:
        method:
          type: string
          minLength: 1
          maxLength: 32
        url:
          type: string
          maxLength: 4096
        headers:
          description: 任意可 JSON 序列化的值
          nullable: true
        body:
          description: 任意可 JSON 序列化的值
          nullable: true
...
    AppAuditError:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          maxLength: 100
        code:
          type: string
          maxLength: 100
        message:
          type: string
          maxLength: 4096
    AppAuditStep:
      type: object
      additionalProperties: false
      required:
        - sequence
        - name
        - startedAt
        - durationMs
      properties:
        sequence:
          type: integer
          minimum: 1
          maximum: 128
        code:
          type: string
          maxLength: 100
        name:
          type: string
          minLength: 1
          maxLength: 200
        startedAt:
          type: string
          format: date-time
        durationMs:
          type: number
          minimum: 0
        status:
          oneOf:
            - type: number
            - type: string
              maxLength: 100
        request:
          $ref: '#/components/schemas/AppAuditRequest'
        response:
          $ref: '#/components/schemas/AppAuditResponse'
        error:
          $ref: '#/components/schemas/AppAuditError'
    AppAudit:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - title
        - metadata
        - steps
      properties:
        schemaVersion:
          type: integer
          enum:
            - 1
        title:
          type: string
          minLength: 1
          maxLength: 200
        metadata:
          type: array
          maxItems: 64
          items:
            $ref: '#/components/schemas/AppAuditMetadata'
        steps:
          type: array
          maxItems: 128
          items:
            $ref: '#/components/schemas/AppAuditStep'
    RequestDetail:
      type: object
      required:
        - id
        - requestId
        - projectName
        - actionName
        - clientId
        - requesterUserId
        - accessTokenId
        - status
        - httpCode
        - latencyMs
        - errorMessage
        - payloadState
        - createdAt
        - finishedAt
        - payloadUnavailable
        - requestPayload
        - responsePayload
        - appAudit
      properties:
        id:
          type: integer
          example: 1
        requestId:
          type: string
          format: uuid
        projectName:
          type: string
          example: cn-nodes
        actionName:
          type: string
          example: ping
        clientId:
          type: string
          nullable: true
        requesterUserId:
          type: integer
          nullable: true
        accessTokenId:
          type: integer
          nullable: true
        status:
          type: string
          example: ok
          description: 业务状态，例如 ok、timeout、error、no_project、no_device、offline、disabled、rejected
        httpCode:
          type: integer
          nullable: true
          example: 200
        latencyMs:
          type: integer
          nullable: true
          minimum: 0
        errorMessage:
          type: string
          nullable: true
        payloadState:
          type: string
          enum:
            - pending
            - indexed
            - failed
            - unavailable
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        payloadUnavailable:
          type: boolean
        requestPayload:
          description: 任意可 JSON 序列化的值
          nullable: true
        responsePayload:
          description: 任意可 JSON 序列化的值
          nullable: true
        appAudit:
          allOf:
            - $ref: '#/components/schemas/AppAudit'
          nullable: true
...
  responses:
    BadRequest:
      description: 请求参数、路径参数或查询参数不符合契约。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 凭证缺失、无效、过期、撤销或对应账号已不存在。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 凭证有效，但账号、功能组或 RBAC 权限不允许本次操作。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 目标资源不存在或已被软删除。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源名称或关联关系与现有数据冲突。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 服务端发生未预期错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'